+++ packages/cosmwasm-launchpad/src/interfaces/index.ts
@@ -1,1 +1,0 @@
-export { Expiration } from "./cw0";